+++ java/compaction/compaction-status-store/src/test/java/sleeper/compaction/status/store/job/StoreCompactionJobExpiryIT.java
@@ -32,7 +32,7 @@ import java.time.temporal.ChronoField;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.core.record.process.status.TestRunStatusUpdates.defaultUpdateTime;
+import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.defaultUpdateTime;
 
 public class StoreCompactionJobExpiryIT extends DynamoDBCompactionJobStatusStoreTestBase {
