+++ java/ingest/ingest-status-store/src/test/java/sleeper/ingest/status/store/job/StoreIngestJobExpiryIT.java
@@ -27,7 +27,7 @@ import java.time.Instant;
 import java.time.temporal.ChronoField;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.core.record.process.status.TestRunStatusUpdates.defaultUpdateTime;
+import static sleeper.core.record.process.status.ProcessStatusUpdateTestHelper.defaultUpdateTime;
 
 public class StoreIngestJobExpiryIT extends DynamoDBIngestJobStatusStoreTestBase {
